modifica stampa

+++ AlgoInC++/codice/StampaVettore.cpp
@@ -10,7 +10,7 @@ void StampaVettore::printVector(const std::vector<uint64_t>& vettore) {
 
 void StampaVettore::printVector(const std::vector<double>& vettore) {
         std::cout << "[ ";
-        for (size_t val : vettore) {
+        for (double val : vettore) {
             std::cout << val << " ";
         }
         std::cout << "]" << std::endl;
